Move core lifecycle events out of repository.events package
