+++ components/nexus-core/src/main/java/org/sonatype/nexus/NxApplication.java
@@ -25,11 +25,11 @@ import org.sonatype.nexus.configuration.ApplicationConfiguration;
 import org.sonatype.nexus.configuration.ApplicationDirectories;
 import org.sonatype.nexus.configuration.ConfigurationChangeEvent;
 import org.sonatype.nexus.events.EventSubscriberHost;
+import org.sonatype.nexus.events.NexusInitializedEvent;
+import org.sonatype.nexus.events.NexusStartedEvent;
+import org.sonatype.nexus.events.NexusStoppedEvent;
+import org.sonatype.nexus.events.NexusStoppingEvent;
 import org.sonatype.nexus.internal.orient.OrientBootstrap;
-import org.sonatype.nexus.proxy.events.NexusInitializedEvent;
-import org.sonatype.nexus.proxy.events.NexusStartedEvent;
-import org.sonatype.nexus.proxy.events.NexusStoppedEvent;
-import org.sonatype.nexus.proxy.events.NexusStoppingEvent;
 import org.sonatype.nexus.proxy.registry.RepositoryRegistry;
 import org.sonatype.nexus.proxy.repository.ShadowRepository;
 import org.sonatype.nexus.scheduling.TaskConfiguration;
