+++ components/nexus-core/src/main/java/org/sonatype/nexus/events/NexusStartedEvent.java
@@ -10,17 +10,17 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.proxy.events;
+package org.sonatype.nexus.events;
 
 /**
- * The event that is occured when nexus has been stopped (fired as first step of shutdown process).
+ * The event that is occured when nexus has started (fired as last step of boot process, everything is in place).
  *
  * @author cstamas
  */
-public class NexusStoppedEvent
+public class NexusStartedEvent
     extends NexusStateChangeEvent
 {
-  public NexusStoppedEvent(Object sender) {
+  public NexusStartedEvent(Object sender) {
     super(sender);
   }
 }
