+++ components/nexus-core/src/main/java/org/sonatype/nexus/events/NexusStoppingEvent.java
@@ -10,26 +10,20 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.proxy.events;
-
-import java.util.Date;
+package org.sonatype.nexus.events;
 
 /**
- * The event that is occured when nexus has started (fired as last step of boot process, everything is in place).
+ * Event fired when Nexus has started to shutdown.
  *
- * @author cstamas
+ * This event is emitted just prior to {@link NexusStoppedEvent} and should be used only
+ * for an early hint that shutdown is going to occur, to stop network connections, etc.
+ *
+ * @since 2.0
  */
-public abstract class NexusStateChangeEvent
-    extends AbstractVetoableEvent<Object>
+public class NexusStoppingEvent
+    extends NexusStateChangeEvent
 {
-  private final Date date;
-
-  public NexusStateChangeEvent(Object sender) {
+  public NexusStoppingEvent(Object sender) {
     super(sender);
-    this.date = new Date();
-  }
-
-  public Date getEventDate() {
-    return date;
   }
 }
