+++ components/nexus-core/src/main/java/org/sonatype/nexus/events/Veto.java
@@ -10,17 +10,28 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.proxy.events;
+package org.sonatype.nexus.events;
 
-/**
- * The event that is occured when nexus has started (fired as last step of boot process, everything is in place).
- *
- * @author cstamas
- */
-public class NexusStartedEvent
-    extends NexusStateChangeEvent
+// FIXME: Remove event veto concept
+
+@Deprecated
+public class Veto
 {
-  public NexusStartedEvent(Object sender) {
-    super(sender);
+  private final Object vetoer;
+
+  private final Throwable reason;
+
+  public Veto(Object vetoer, Throwable reason) {
+    this.vetoer = vetoer;
+
+    this.reason = reason;
+  }
+
+  public Object getVetoer() {
+    return vetoer;
+  }
+
+  public Throwable getReason() {
+    return reason;
   }
 }
