+++ plugins/capabilities/nexus-capabilities-plugin/src/main/java/org/sonatype/nexus/capability/internal/CapabilityRegistryBooter.java
@@ -19,8 +19,8 @@ import javax.inject.Singleton;
 import org.sonatype.nexus.capability.CapabilityRegistryEvent.Ready;
 import org.sonatype.nexus.capability.internal.storage.OrientCapabilityStorage;
 import org.sonatype.nexus.events.EventSubscriber;
-import org.sonatype.nexus.proxy.events.NexusInitializedEvent;
-import org.sonatype.nexus.proxy.events.NexusStoppingEvent;
+import org.sonatype.nexus.events.NexusInitializedEvent;
+import org.sonatype.nexus.events.NexusStoppingEvent;
 import org.sonatype.sisu.goodies.eventbus.EventBus;
 
 import com.google.common.base.Throwables;
