+++ plugins/capabilities/nexus-capabilities-plugin/src/main/java/org/sonatype/nexus/capability/internal/condition/NexusIsActiveCondition.java
@@ -18,8 +18,8 @@ import javax.inject.Provider;
 
 import org.sonatype.nexus.capability.Condition;
 import org.sonatype.nexus.capability.support.condition.ConditionSupport;
-import org.sonatype.nexus.proxy.events.NexusStartedEvent;
-import org.sonatype.nexus.proxy.events.NexusStoppedEvent;
+import org.sonatype.nexus.events.NexusStartedEvent;
+import org.sonatype.nexus.events.NexusStoppedEvent;
 import org.sonatype.sisu.goodies.eventbus.EventBus;
 
 import com.google.common.eventbus.Subscribe;
