+++ plugins/capabilities/nexus-capabilities-plugin/src/test/java/org/sonatype/nexus/capability/internal/DefaultCapabilityReferenceTest.java
@@ -29,8 +29,8 @@ import org.sonatype.nexus.capability.support.condition.CryptoConditions;
 import org.sonatype.nexus.capability.support.condition.LogicalConditions;
 import org.sonatype.nexus.capability.support.condition.NexusConditions;
 import org.sonatype.nexus.capability.support.condition.RepositoryConditions;
-import org.sonatype.nexus.proxy.events.NexusStartedEvent;
-import org.sonatype.nexus.proxy.events.NexusStoppedEvent;
+import org.sonatype.nexus.events.NexusStartedEvent;
+import org.sonatype.nexus.events.NexusStoppedEvent;
 import org.sonatype.sisu.goodies.eventbus.EventBus;
 import org.sonatype.sisu.goodies.eventbus.internal.DefaultEventBus;
 import org.sonatype.sisu.litmus.testsupport.TestSupport;
